+++ src/models/documents/NetpbmGraphic/parser/rasterParser.tsx
