+++ src/models/documents/NetpbmGraphic/types.d.ts
